requirements.txt
python-dotenv
pandas 
numpy 
pymongo
certifi 
scikit-learn
scipy
dill
pyaml


#Add at the last -e .